Extract createBackdropSlice from projectStore (P5 store slice)
